+++ CHANGES.md
@@ -5,6 +5,8 @@
 - [issue #41] Support `MANTA_KEY_ID=SHA256:...` fingerprint form.
 - [issue #43] Ensure a broken `~/.ssh/*.pub` symlink doesn't break
   finding a key for a given `MANTA_KEY_ID`.
+- [issue #42] Support `~~` and `~~/` aliases for `/USER` a la
+  node-manta.
 
 
 ## 2.6.0